Updates on buttons and sound
First of all: I apologise to anyone who actually reads my commits and the commentary. First, I thought great I can commit and commit and have a documentation of every step I took, but I soon realised, that this might be a bit too much. So from now on, I'll do more general commits with lots of changes. So here it goes:

- Removed Section button as I don't need it
- Added bird sounds for the viewers convenience
- Designed the title a bit, but this is a neverending story