Externalised Textboxes, added possibility for language selection
So today I wanted to externalise the texts from the HTML, so translations can be integrated more easiliy. I creates therefore a JSON file for the English text and included it in the HTML file via JavaScript. Translations can be added by adding new JSON files and switching the language on the website (will follow).